+++ Models/Responses/BacklinksBulkBacklinksLiveResultInfo.cs
@@ -9,7 +9,7 @@ namespace DataForSeo.Client.Models.Responses
         /// the number of results returned in the items array
         /// </summary>
         [Newtonsoft.Json.JsonProperty("items_count", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public long? ItemsCount { get; set; }
+        public double? ItemsCount { get; set; }
 
         /// <summary>
         /// contains relevant backlink data
