+++ Models/Responses/BusinessDataBusinessListingsLocationsResultInfo.cs
@@ -21,7 +21,7 @@ namespace DataForSeo.Client.Models.Responses
         /// number of businesses in this location in our database
         /// </summary>
         [Newtonsoft.Json.JsonProperty("business_count", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public long? BusinessCount { get; set; }
+        public double? BusinessCount { get; set; }
 
         private System.Collections.Generic.IDictionary<string, object> _additionalProperties;
         [Newtonsoft.Json.JsonExtensionData]
