+++ Models/Responses/DataforseoLabsGoogleCategoriesForKeywordsLiveResultInfo.cs
@@ -16,7 +16,7 @@ namespace DataForSeo.Client.Models.Responses
         /// the number of results returned in the items array
         /// </summary>
         [Newtonsoft.Json.JsonProperty("items_count", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public long? ItemsCount { get; set; }
+        public double? ItemsCount { get; set; }
 
         /// <summary>
         /// contains keywords and related keyword difficulty scores
